Move table code into dir
