+++ src/components/clientDashboard/AuditHistory.tsx
@@ -4,7 +4,7 @@ import { capitalize, isNil } from 'lodash-es';
 
 import SimpleTable from '../elements/SimpleTable';
 
-import { ColumnDef } from '@/components/elements/GenericTable';
+import { ColumnDef } from '@/components/elements/table/types';
 import useSafeParams from '@/hooks/useSafeParams';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import HmisField from '@/modules/hmis/components/HmisField';
