+++ src/components/clientDashboard/enrollments/tabs/AssessmentsTable.tsx
@@ -1,8 +1,8 @@
 import { startCase } from 'lodash-es';
 import { useCallback } from 'react';
 
-import AssessmentStatus from '@/components/elements/AssessmentStatus';
-import { ColumnDef } from '@/components/elements/GenericTable';
+import { ColumnDef } from '@/components/elements/table/types';
+import AssessmentStatus from '@/modules/assessments/components/AssessmentStatus';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import {
   parseAndFormatDate,
