+++ src/components/clientDashboard/enrollments/tabs/EventsTable.tsx
@@ -1,4 +1,4 @@
-import { ColumnDef } from '@/components/elements/GenericTable';
+import { ColumnDef } from '@/components/elements/table/types';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import HmisEnum from '@/modules/hmis/components/HmisEnum';
 import {
