+++ src/modules/household/components/QuickAddHouseholdMembers.tsx
@@ -12,7 +12,7 @@ import { RecentHouseholdMember } from '../types';
 import AssociatedHouseholdMembers from './AssociatedHouseholdMembers';
 import RelationshipToHohSelect from './RelationshipToHohSelect';
 
-import { ColumnDef } from '@/components/elements/GenericTable';
+import { ColumnDef } from '@/components/elements/table/types';
 import { clientBriefName } from '@/modules/hmis/hmisUtil';
 import { RelationshipToHoH } from '@/types/gqlTypes';
