+++ src/modules/projects/components/Funders.tsx
@@ -2,7 +2,7 @@ import AddIcon from '@mui/icons-material/Add';
 import { Paper } from '@mui/material';
 
 import ButtonLink from '@/components/elements/ButtonLink';
-import { ColumnDef } from '@/components/elements/GenericTable';
+import { ColumnDef } from '@/components/elements/table/types';
 import PageTitle from '@/components/layout/PageTitle';
 import useSafeParams from '@/hooks/useSafeParams';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
