+++ src/modules/projects/components/tables/ProjectReferralRequestsTable.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 
-import { ColumnDef } from '@/components/elements/GenericTable';
+import { ColumnDef } from '@/components/elements/table/types';
 import DeleteMutationButton from '@/modules/dataFetching/components/DeleteMutationButton';
 import GenericTableWithData from '@/modules/dataFetching/components/GenericTableWithData';
 import {
